Set update state in fu_engine_write_firmware() for failures
If the update fails, set the failure before calling cleanup so that
a plugin can potentially skip cleanup steps in a failure case.

(cherry picked from commit c6803f1a1295cc982d39677d627f5b36837dab5d)

+++ plugins/redfish/fu-redfish-plugin.c
@@ -559,6 +559,10 @@ fu_redfish_plugin_cleanup(FuPlugin *plugin,
 	if (!fu_device_has_private_flag(device, FU_REDFISH_DEVICE_FLAG_MANAGER_RESET))
 		return TRUE;
 
+	/* update failed; don't reboot BMC */
+	if (fu_device_get_update_state(device) == FWUPD_UPDATE_STATE_FAILED)
+		return TRUE;
+
 	/* progress */
 	fu_progress_set_id(progress, G_STRLOC);
 	fu_progress_add_step(progress, FWUPD_STATUS_DEVICE_BUSY, 1, "manager-reboot");
